Add regionEquivalence: a venn diagram representation of objXprop
regionEquivalence takes the 8-state multinomial as probabilities
over the regions of the venn diagram. For each region, a coin is
flipped weighted by the probability of that region. This probability
varies as a function of `n_objects`, by `1 - (1-theta)^n`. That is,
the probability of that type of property existing across many
objects = 1 - probability of the property not existing across all
of the objects.

This computation takes around 50ms to run, as opposed to ~500ms of
using the objects with properties representation with 4 objects.